requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.4
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-celery-beat==2.0.0
django-comments-xtd==2.6.1
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-haystack==2.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mptt==0.11.0
django-pagedown==2.1.2
django-taggit==1.3.0
django-timezone-field==4.0
djangorestframework==3.11.0
docutils==0.16
flower==0.9.4
future==0.18.2
gunicorn==20.0.4
humanize==0.5.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
markdown2==2.3.9
Pillow==7.1.2
psycopg2==2.8.5
pysolr==3.9.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
